Fix hadoop-queue checkstyle
