+++ eagle-jpm/eagle-hadoop-queue/src/main/java/org/apache/eagle/hadoop/queue/crawler/RunningAppsCrawler.java
@@ -18,11 +18,11 @@
 
 package org.apache.eagle.hadoop.queue.crawler;
 
-import backtype.storm.spout.SpoutOutputCollector;
-import org.apache.eagle.hadoop.queue.common.YarnClusterResourceURLBuilder;
-import org.apache.eagle.hadoop.queue.model.applications.Apps;
 import org.apache.eagle.hadoop.queue.common.HadoopYarnResourceUtils;
+import org.apache.eagle.hadoop.queue.common.YarnClusterResourceURLBuilder;
 import org.apache.eagle.hadoop.queue.model.applications.AppsWrapper;
+
+import backtype.storm.spout.SpoutOutputCollector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -30,37 +30,37 @@ import java.io.IOException;
 
 public class RunningAppsCrawler implements Runnable {
 
-	private final static Logger logger = LoggerFactory.getLogger(RunningAppsCrawler.class);
+    private static final Logger logger = LoggerFactory.getLogger(RunningAppsCrawler.class);
 
-	private RunningAppParseListener listener;
-	private String urlString;
+    private RunningAppParseListener listener;
+    private String urlString;
 
-	public RunningAppsCrawler(String site, String baseUrl, SpoutOutputCollector collector){
-		this.urlString = YarnClusterResourceURLBuilder.buildRunningAppsURL(baseUrl);
-		//this.urlString = YarnClusterResourceURLBuilder.buildFinishedAppsURL(baseUrl);
-		listener = new RunningAppParseListener(site, collector);
-	}
+    public RunningAppsCrawler(String site, String baseUrl, SpoutOutputCollector collector) {
+        this.urlString = YarnClusterResourceURLBuilder.buildRunningAppsURL(baseUrl);
+        //this.urlString = YarnClusterResourceURLBuilder.buildFinishedAppsURL(baseUrl);
+        listener = new RunningAppParseListener(site, collector);
+    }
 
-	@Override
-	public void run() {
-		try {
-			logger.info("Start to crawl app metrics from " + this.urlString);
-			AppsWrapper appsWrapper = (AppsWrapper) HadoopYarnResourceUtils.getObjectFromStreamWithGzip(urlString, AppsWrapper.class);
-			if(appsWrapper == null || appsWrapper.getApps() == null) {
-				logger.error("Failed to crawl running applications with api = " + urlString);
-			} else {
-				long currentTimestamp = System.currentTimeMillis();
-				listener.onMetric(appsWrapper.getApps(), currentTimestamp);
-			}
-		} catch (IOException e) {
-			logger.error(e.getMessage());
-			if(logger.isDebugEnabled()) {
-				logger.trace(e.getMessage(), e);
-			}
-		} catch (Exception e) {
-			logger.error(e.getMessage(), e);
-		} finally {
-			listener.flush();
-		}
-	}
+    @Override
+    public void run() {
+        try {
+            logger.info("Start to crawl app metrics from " + this.urlString);
+            AppsWrapper appsWrapper = (AppsWrapper) HadoopYarnResourceUtils.getObjectFromStreamWithGzip(urlString, AppsWrapper.class);
+            if (appsWrapper == null || appsWrapper.getApps() == null) {
+                logger.error("Failed to crawl running applications with api = " + urlString);
+            } else {
+                long currentTimestamp = System.currentTimeMillis();
+                listener.onMetric(appsWrapper.getApps(), currentTimestamp);
+            }
+        } catch (IOException e) {
+            logger.error(e.getMessage());
+            if (logger.isDebugEnabled()) {
+                logger.trace(e.getMessage(), e);
+            }
+        } catch (Exception e) {
+            logger.error(e.getMessage(), e);
+        } finally {
+            listener.flush();
+        }
+    }
 }
