+++ eagle-jpm/eagle-hadoop-queue/src/main/java/org/apache/eagle/hadoop/queue/crawler/SchedulerInfoCrawler.java
@@ -18,10 +18,11 @@
 
 package org.apache.eagle.hadoop.queue.crawler;
 
-import backtype.storm.spout.SpoutOutputCollector;
-import org.apache.eagle.hadoop.queue.common.YarnClusterResourceURLBuilder;
-import org.apache.eagle.hadoop.queue.model.scheduler.*;
 import org.apache.eagle.hadoop.queue.common.HadoopYarnResourceUtils;
+import org.apache.eagle.hadoop.queue.common.YarnClusterResourceURLBuilder;
+import org.apache.eagle.hadoop.queue.model.scheduler.SchedulerInfo;
+import org.apache.eagle.hadoop.queue.model.scheduler.SchedulerWrapper;
+import backtype.storm.spout.SpoutOutputCollector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -29,38 +30,38 @@ import java.io.IOException;
 
 public class SchedulerInfoCrawler implements Runnable {
 
-	private final static Logger logger = LoggerFactory.getLogger(SchedulerInfoCrawler.class);
+    private static final Logger logger = LoggerFactory.getLogger(SchedulerInfoCrawler.class);
 
-	private SchedulerInfoParseListener listener;
+    private SchedulerInfoParseListener listener;
     private String urlString;
 
-	public SchedulerInfoCrawler(String site, String baseUrl, SpoutOutputCollector collector) {
-		this.urlString = YarnClusterResourceURLBuilder.buildSchedulerInfoURL(baseUrl);
-		this.listener = new SchedulerInfoParseListener(site, collector);
-	}
+    public SchedulerInfoCrawler(String site, String baseUrl, SpoutOutputCollector collector) {
+        this.urlString = YarnClusterResourceURLBuilder.buildSchedulerInfoURL(baseUrl);
+        this.listener = new SchedulerInfoParseListener(site, collector);
+    }
 
-	@Override
-	public void run() {
-		try {
-			//https://some.server.address:50030/ws/v1/cluster/scheduler?anonymous=true
-			logger.info("Start to crawl cluster scheduler queues from " + this.urlString);
-			SchedulerWrapper schedulerWrapper = (SchedulerWrapper) HadoopYarnResourceUtils.getObjectFromStreamWithGzip(urlString, SchedulerWrapper.class);
-			if (schedulerWrapper == null || schedulerWrapper.getScheduler() == null) {
-				logger.error("Failed to crawl scheduler info with url = " + this.urlString);
-			} else {
-				SchedulerInfo scheduler = schedulerWrapper.getScheduler().getSchedulerInfo();
-				logger.info("Crawled " + scheduler.getQueues().getQueue().size() + " queues");
-				long currentTimestamp = System.currentTimeMillis();
-				listener.onMetric(scheduler, currentTimestamp);
-			}
-		} catch (IOException e) {
-			logger.error("Got IO exception while connecting to "+this.urlString + " : "+ e.getMessage());
-		} catch (Exception e) {
-			logger.error("Got exception while crawling queues:" + e.getMessage(), e);
-		} catch (Throwable e) {
-			logger.error("Got throwable exception while crawling queues:" + e.getMessage(), e);
-		} finally {
-			listener.flush();
-		}
-	}
+    @Override
+    public void run() {
+        try {
+            //https://some.server.address:50030/ws/v1/cluster/scheduler?anonymous=true
+            logger.info("Start to crawl cluster scheduler queues from " + this.urlString);
+            SchedulerWrapper schedulerWrapper = (SchedulerWrapper) HadoopYarnResourceUtils.getObjectFromStreamWithGzip(urlString, SchedulerWrapper.class);
+            if (schedulerWrapper == null || schedulerWrapper.getScheduler() == null) {
+                logger.error("Failed to crawl scheduler info with url = " + this.urlString);
+            } else {
+                SchedulerInfo scheduler = schedulerWrapper.getScheduler().getSchedulerInfo();
+                logger.info("Crawled " + scheduler.getQueues().getQueue().size() + " queues");
+                long currentTimestamp = System.currentTimeMillis();
+                listener.onMetric(scheduler, currentTimestamp);
+            }
+        } catch (IOException e) {
+            logger.error("Got IO exception while connecting to " + this.urlString + " : " + e.getMessage());
+        } catch (Exception e) {
+            logger.error("Got exception while crawling queues:" + e.getMessage(), e);
+        } catch (Throwable e) {
+            logger.error("Got throwable exception while crawling queues:" + e.getMessage(), e);
+        } finally {
+            listener.flush();
+        }
+    }
 }
