+++ eagle-jpm/eagle-hadoop-queue/src/main/java/org/apache/eagle/hadoop/queue/crawler/SchedulerInfoParseListener.java
@@ -18,23 +18,26 @@
 
 package org.apache.eagle.hadoop.queue.crawler;
 
-import backtype.storm.spout.SpoutOutputCollector;
 import org.apache.eagle.dataproc.impl.storm.ValuesArray;
 import org.apache.eagle.hadoop.queue.common.HadoopClusterConstants;
 import org.apache.eagle.hadoop.queue.common.HadoopClusterConstants.MetricName;
 import org.apache.eagle.hadoop.queue.model.scheduler.*;
-import org.apache.eagle.hadoop.queue.model.scheduler.Queue;
 import org.apache.eagle.hadoop.queue.storm.HadoopQueueMessageId;
 import org.apache.eagle.log.entity.GenericMetricEntity;
+
+import backtype.storm.spout.SpoutOutputCollector;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 
 
 public class SchedulerInfoParseListener {
 
-    private final static Logger LOG = LoggerFactory.getLogger(SchedulerInfoParseListener.class);
+    private static final Logger LOG = LoggerFactory.getLogger(SchedulerInfoParseListener.class);
     //private final static long AGGREGATE_INTERVAL = DateTimeUtil.ONEMINUTE;
     //private int MAX_CACHE_COUNT = 1000;
 
@@ -50,10 +53,10 @@ public class SchedulerInfoParseListener {
     }
 
     public void onMetric(SchedulerInfo scheduler, long currentTimestamp) throws Exception {
-        Map<String,String> tags = buildMetricTags(null, null);
+        Map<String, String> tags = buildMetricTags(null, null);
         createMetric(MetricName.HADOOP_CLUSTER_CAPACITY, tags, currentTimestamp, scheduler.getCapacity());
         createMetric(MetricName.HADOOP_CLUSTER_USED_CAPACITY, tags, currentTimestamp, scheduler.getUsedCapacity());
-        for(Queue queue : scheduler.getQueues().getQueue()) {
+        for (Queue queue : scheduler.getQueues().getQueue()) {
             createQueues(queue, currentTimestamp, scheduler, null);
         }
     }
@@ -74,7 +77,7 @@ public class SchedulerInfoParseListener {
     }
 
     private Map<String, String> buildMetricTags(String queueName, String parentQueueName) {
-        Map<String,String> tags = new HashMap<>();
+        Map<String, String> tags = new HashMap<>();
         tags.put(HadoopClusterConstants.TAG_SITE, this.site);
         if (queueName != null) {
             tags.put(HadoopClusterConstants.TAG_QUEUE, queueName);
@@ -85,12 +88,12 @@ public class SchedulerInfoParseListener {
         return tags;
     }
 
-    private void createMetric(String metricName, Map<String,String> tags,long timestamp, double value) throws Exception {
+    private void createMetric(String metricName, Map<String, String> tags, long timestamp, double value) throws Exception {
         GenericMetricEntity e = new GenericMetricEntity();
         e.setPrefix(metricName);
         e.setTimestamp(timestamp);
         e.setTags(tags);
-        e.setValue(new double[]{value});
+        e.setValue(new double[] {value});
         this.metricEntities.add(e);
     }
 
@@ -111,7 +114,7 @@ public class SchedulerInfoParseListener {
         _entity.setTimestamp(currentTimestamp);
 
         List<UserWrapper> userList = new ArrayList<>();
-        if (queue.getUsers() != null && queue.getUsers().getUser() != null )  {
+        if (queue.getUsers() != null && queue.getUsers().getUser() != null) {
             for (User user : queue.getUsers().getUser()) {
                 UserWrapper newUser = new UserWrapper(user);
                 userList.add(newUser);
@@ -122,20 +125,21 @@ public class SchedulerInfoParseListener {
         runningQueueAPIEntities.add(_entity);
 
         createMetric(MetricName.HADOOP_QUEUE_NUMPENDING_JOBS, _tags, currentTimestamp, queue.getNumPendingApplications());
-        createMetric(MetricName.HADOOP_QUEUE_USED_CAPACITY,_tags,currentTimestamp,queue.getAbsoluteUsedCapacity());
-        if(queue.getAbsoluteCapacity() == 0) {
-            createMetric(MetricName.HADOOP_QUEUE_USED_CAPACITY_RATIO,_tags,currentTimestamp,0);
+        createMetric(MetricName.HADOOP_QUEUE_USED_CAPACITY, _tags, currentTimestamp, queue.getAbsoluteUsedCapacity());
+        if (queue.getAbsoluteCapacity() == 0) {
+            createMetric(MetricName.HADOOP_QUEUE_USED_CAPACITY_RATIO, _tags, currentTimestamp, 0);
         } else {
-            createMetric(MetricName.HADOOP_QUEUE_USED_CAPACITY_RATIO,_tags,currentTimestamp,queue.getAbsoluteUsedCapacity()/queue.getAbsoluteCapacity());
+            createMetric(MetricName.HADOOP_QUEUE_USED_CAPACITY_RATIO, _tags, currentTimestamp, queue.getAbsoluteUsedCapacity() / queue.getAbsoluteCapacity());
         }
 
-        if (queue.getUsers() != null && queue.getUsers().getUser() != null )  {
+        if (queue.getUsers() != null && queue.getUsers().getUser() != null) {
             for (User user : queue.getUsers().getUser()) {
                 Map<String, String> userTags = new HashMap<>(_tags);
                 userTags.put(HadoopClusterConstants.TAG_USER, user.getUsername());
-                createMetric(HadoopClusterConstants.MetricName.HADOOP_USER_NUMPENDING_JOBS,userTags,currentTimestamp,user.getNumPendingApplications());
-                createMetric(HadoopClusterConstants.MetricName.HADOOP_USER_USED_MEMORY,userTags,currentTimestamp,user.getResourcesUsed().getMemory());
-                createMetric(HadoopClusterConstants.MetricName.HADOOP_USER_USED_MEMORY_RATIO,userTags,currentTimestamp,((double)user.getResourcesUsed().getMemory()) / queue.getResourcesUsed().getMemory());
+                createMetric(HadoopClusterConstants.MetricName.HADOOP_USER_NUMPENDING_JOBS, userTags, currentTimestamp, user.getNumPendingApplications());
+                createMetric(HadoopClusterConstants.MetricName.HADOOP_USER_USED_MEMORY, userTags, currentTimestamp, user.getResourcesUsed().getMemory());
+                createMetric(HadoopClusterConstants.MetricName.HADOOP_USER_USED_MEMORY_RATIO, userTags, currentTimestamp,
+                    ((double) user.getResourcesUsed().getMemory()) / queue.getResourcesUsed().getMemory());
             }
         }
