+++ eagle-jpm/eagle-hadoop-queue/src/main/java/org/apache/eagle/hadoop/queue/model/scheduler/RunningQueueAPIEntity.java
@@ -25,134 +25,134 @@ import org.codehaus.jackson.map.annotate.JsonSerialize;
 
 import java.util.List;
 
-@JsonSerialize(include=JsonSerialize.Inclusion.NON_NULL)
+@JsonSerialize(include = JsonSerialize.Inclusion.NON_NULL)
 @Table("running_queue")
 @ColumnFamily("f")
 @Prefix("rqueue")
 @Service(HadoopClusterConstants.RUNNING_QUEUE_SERVICE_NAME)
 @TimeSeries(true)
-@Partition({"site"})
+@Partition( {"site"})
 public class RunningQueueAPIEntity extends TaggedLogAPIEntity {
-	@Column("a")
-	private String state;
-	@Column("b")
-	private double absoluteCapacity;
-	@Column("c")
-	private double absoluteMaxCapacity;
-	@Column("d")
-	private double absoluteUsedCapacity;
-	@Column("e")
-	private long memory;
-	@Column("f")
-	private long vcores;
-	@Column("g")
-	private int numActiveApplications;
-	@Column("h")
-	private int numPendingApplications;
-	@Column("i")
-	private int maxActiveApplications;
-	@Column("j")
-	private String scheduler;
-	@Column("k")
-	private List<UserWrapper> users;
+    @Column("a")
+    private String state;
+    @Column("b")
+    private double absoluteCapacity;
+    @Column("c")
+    private double absoluteMaxCapacity;
+    @Column("d")
+    private double absoluteUsedCapacity;
+    @Column("e")
+    private long memory;
+    @Column("f")
+    private long vcores;
+    @Column("g")
+    private int numActiveApplications;
+    @Column("h")
+    private int numPendingApplications;
+    @Column("i")
+    private int maxActiveApplications;
+    @Column("j")
+    private String scheduler;
+    @Column("k")
+    private List<UserWrapper> users;
 
-	public String getScheduler() {
-		return scheduler;
-	}
+    public String getScheduler() {
+        return scheduler;
+    }
 
-	public void setScheduler(String scheduler) {
-		this.scheduler = scheduler;
-		valueChanged("scheduler");
-	}
+    public void setScheduler(String scheduler) {
+        this.scheduler = scheduler;
+        valueChanged("scheduler");
+    }
 
-	public int getMaxActiveApplications() {
-		return maxActiveApplications;
-	}
+    public int getMaxActiveApplications() {
+        return maxActiveApplications;
+    }
 
-	public void setMaxActiveApplications(int maxActiveApplications) {
-		this.maxActiveApplications = maxActiveApplications;
-		valueChanged("maxActiveApplications");
-	}
+    public void setMaxActiveApplications(int maxActiveApplications) {
+        this.maxActiveApplications = maxActiveApplications;
+        valueChanged("maxActiveApplications");
+    }
 
-	public String getState() {
-		return state;
-	}
+    public String getState() {
+        return state;
+    }
 
-	public void setState(String state) {
-		this.state = state;
-		valueChanged("state");
+    public void setState(String state) {
+        this.state = state;
+        valueChanged("state");
 
-	}
+    }
 
-	public double getAbsoluteCapacity() {
-		return absoluteCapacity;
-	}
+    public double getAbsoluteCapacity() {
+        return absoluteCapacity;
+    }
 
-	public void setAbsoluteCapacity(double absoluteCapacity) {
-		this.absoluteCapacity = absoluteCapacity;
-		valueChanged("absoluteCapacity");
-	}
+    public void setAbsoluteCapacity(double absoluteCapacity) {
+        this.absoluteCapacity = absoluteCapacity;
+        valueChanged("absoluteCapacity");
+    }
 
-	public double getAbsoluteMaxCapacity() {
-		return absoluteMaxCapacity;
-	}
+    public double getAbsoluteMaxCapacity() {
+        return absoluteMaxCapacity;
+    }
 
-	public void setAbsoluteMaxCapacity(double absoluteMaxCapacity) {
-		this.absoluteMaxCapacity = absoluteMaxCapacity;
-		valueChanged("absoluteMaxCapacity");
-	}
+    public void setAbsoluteMaxCapacity(double absoluteMaxCapacity) {
+        this.absoluteMaxCapacity = absoluteMaxCapacity;
+        valueChanged("absoluteMaxCapacity");
+    }
 
-	public double getAbsoluteUsedCapacity() {
-		return absoluteUsedCapacity;
-	}
+    public double getAbsoluteUsedCapacity() {
+        return absoluteUsedCapacity;
+    }
 
-	public void setAbsoluteUsedCapacity(double absoluteUsedCapacity) {
-		this.absoluteUsedCapacity = absoluteUsedCapacity;
-		valueChanged("absoluteUsedCapacity");
-	}
+    public void setAbsoluteUsedCapacity(double absoluteUsedCapacity) {
+        this.absoluteUsedCapacity = absoluteUsedCapacity;
+        valueChanged("absoluteUsedCapacity");
+    }
 
-	public long getMemory() {
-		return memory;
-	}
+    public long getMemory() {
+        return memory;
+    }
 
-	public void setMemory(long memory) {
-		this.memory = memory;
-		valueChanged("memory");
-	}
+    public void setMemory(long memory) {
+        this.memory = memory;
+        valueChanged("memory");
+    }
 
-	public long getVcores() {
-		return vcores;
-	}
+    public long getVcores() {
+        return vcores;
+    }
 
-	public void setVcores(long vcores) {
-		this.vcores = vcores;
-		valueChanged("vcores");
-	}
+    public void setVcores(long vcores) {
+        this.vcores = vcores;
+        valueChanged("vcores");
+    }
 
-	public int getNumActiveApplications() {
-		return numActiveApplications;
-	}
+    public int getNumActiveApplications() {
+        return numActiveApplications;
+    }
 
-	public void setNumActiveApplications(int numActiveApplications) {
-		this.numActiveApplications = numActiveApplications;
-		valueChanged("numActiveApplications");
-	}
+    public void setNumActiveApplications(int numActiveApplications) {
+        this.numActiveApplications = numActiveApplications;
+        valueChanged("numActiveApplications");
+    }
 
-	public int getNumPendingApplications() {
-		return numPendingApplications;
-	}
+    public int getNumPendingApplications() {
+        return numPendingApplications;
+    }
 
-	public void setNumPendingApplications(int numPendingApplications) {
-		this.numPendingApplications = numPendingApplications;
-		valueChanged("numPendingApplications");
-	}
+    public void setNumPendingApplications(int numPendingApplications) {
+        this.numPendingApplications = numPendingApplications;
+        valueChanged("numPendingApplications");
+    }
 
-	public List<UserWrapper> getUsers() {
-		return users;
-	}
+    public List<UserWrapper> getUsers() {
+        return users;
+    }
 
-	public void setUsers(List<UserWrapper> users) {
-		this.users = users;
-		valueChanged("numPendingApplications");
-	}
+    public void setUsers(List<UserWrapper> users) {
+        this.users = users;
+        valueChanged("numPendingApplications");
+    }
 }
